Edit ExtractorLiveLib logic
