+++ ExtractorProject/Extractors/ExtractorLiveLib.cs
@@ -1,7 +1,22 @@
+using ExtractorService.Extractor;
 using ExtractorService.Models;
 
-namespace ExtractorService.Extractor{
-    public class ExtractorLiveLib<T> : AbstractExtractor, IExtractor<T>{
-        
+namespace ExtractorProject.Extractors
+{
+    public class ExtractorLiveLib<T> : AbstractExtractor, IExtractor<T>
+    {
+        public ExtractorLiveLib(ExtractorSettings Settings) : base(Settings)
+        {
+        }
+
+        public ExtractBatchResult<T> ExtractNextBatch()
+        {
+            throw new NotImplementedException();
+        }
+
+        public bool IsEndData()
+        {
+            throw new NotImplementedException();
+        }
     }
 }
