+++ ExtractorProject/Extractors/IExtractorLiveLib.cs
@@ -1,0 +1,10 @@
+﻿using ExtractorService.Models;
+
+namespace ExtractorService.Extractor
+{
+    public interface IExtractorLiveLib<T>
+    {
+        ExtractBatchResult<T> ExtractNextBatch();
+        bool IsEndData();
+    }
+}
